app/requirements.txt
numpy==1.21
matplotlib==3.0.2
mysql-connector-python==8.0.15
PyMySQL==0.9.3
pytz==2018.9
requests==2.27.1
schedule==0.6.0
six==1.12.0
SQLAlchemy==1.3.0
urllib3==1.26.5
wrapt==1.11.1
pandas==0.25.3